+++ crates/goose-providers/src/canonical/mod.rs
@@ -1,3 +1,4 @@
+pub mod catalog;
 mod model;
 mod name_builder;
 mod registry;
